ISSUE #181 Complete. Refactored ScrumbledData into multiple files
